Basic manipulating of values implemented
It is possible to amnipulate values in the text as defined in #12.
However, this is only implemented for one value (body text). This commit
is a basic version of the features demanded in #11.

For this to work properly (the values to be changeable seperately), the
calc() function needs to be rewritten, which is a whole ather issue.